+++ src/main/scala/mqtt/parser/Monad.scala
@@ -7,9 +7,6 @@ trait Monad[F[_]] {
 }
 
 object Monad {
-  
-  def map2[F[_] : Monad, A, B, C](ma: F[A])(mb: => F[B])(f: (A, B) => C): F[C] = ma.flatMap(a => mb.map(b => f(a, b)))
-  
   implicit class RichMonad[F[_] : Monad, A](ma: F[A]) {
     val monad = implicitly[Monad[F]]
     
@@ -17,6 +14,5 @@ object Monad {
     
     def map[B](f: A => B): F[B] = ma.flatMap(a => monad.unit(f(a)))
   }
-  
 }
