+++ src/main/scala/mqtt/parser/MqttPacketParser.scala
@@ -3,8 +3,7 @@ package mqtt.parser
 import mqtt.PacketParser
 import mqtt.model.ErrorPacket.MalformedPacket
 import mqtt.model.Packet
-import mqtt.parser.Parsers.MqttParser._
-
+import mqtt.parser.MqttPacketsParsers._
 import mqtt.utils.Bit
 
 object MqttPacketParser extends PacketParser {
