+++ src/test/scala/mqtt/parser/MqttFragmentsParsersTest.scala
@@ -1,11 +1,12 @@
 package mqtt.parser
 
+import mqtt.model.Packet.ConnectReturnCode._
 import mqtt.model.QoS
-import mqtt.parser.Parsers.MqttParser._
+import mqtt.parser.MqttFragmentsParsers._
 import org.scalatest.{FunSuite, Matchers}
 import mqtt.utils.BitImplicits._
 
-class FragmentParserTest extends FunSuite with Matchers {
+class MqttFragmentsParsersTest extends FunSuite with Matchers {
   //Qos parser
   test("A QoS parser SHOULD parse a QoS 0") {
     assert(qos().run(Seq(0, 0)) == List((QoS(0), Seq())))
@@ -38,5 +39,28 @@ class FragmentParserTest extends FunSuite with Matchers {
   }
   test("A WillFlags parser SHOULD NOT parse if will_flag=0, QoS=2 [MQTT-3.1.2-13]") {
     assert(willFlags().run(Seq(0, 1, 0, 0)) == List())
+  }
+  
+  //ConnectReturnCode
+  test("A ConnectReturnCode parser SHOULD parse a ConnectionAccepted") {
+    assert(connectReturnCode().run(0.toByte.bits) == List((ConnectionAccepted, Seq())))
+  }
+  test("A ConnectReturnCode parser SHOULD parse a UnacceptableProtocolVersion") {
+    assert(connectReturnCode().run(1.toByte.bits) == List((UnacceptableProtocolVersion, Seq())))
+  }
+  test("A ConnectReturnCode parser SHOULD parse a IdentifierRejected") {
+    assert(connectReturnCode().run(2.toByte.bits) == List((IdentifierRejected, Seq())))
+  }
+  test("A ConnectReturnCode parser SHOULD parse a ServerUnavailable") {
+    assert(connectReturnCode().run(3.toByte.bits) == List((ServerUnavailable, Seq())))
+  }
+  test("A ConnectReturnCode parser SHOULD parse a BadUsernameOrPassword") {
+    assert(connectReturnCode().run(4.toByte.bits) == List((BadUsernameOrPassword, Seq())))
+  }
+  test("A ConnectReturnCode parser SHOULD parse a NotAuthorized") {
+    assert(connectReturnCode().run(5.toByte.bits) == List((NotAuthorized, Seq())))
+  }
+  test("A ConnectReturnCode parser SHOULD parse a byte>5") {
+    assert(connectReturnCode().run(31.toByte.bits) == List())
   }
 }
